requirements.txt
Pillow
pyheif
tk